+++ src/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import './style/App.scss';
 import Navigation from './Navigation';
-import Home from './views/pages/Home';
+import Home from './views/pages/Home.jsx';
 import Find from './views/pages/Find';
 import Modal from './views/pages/Modal';
 import { BrowserRouter as Router, Route } from "react-router-dom";
